requirements.txt
# Rider-PC Client Requirements
# Python 3.9+

# Tooling
pip==26.0

# Web Framework
fastapi==0.115.5
uvicorn[standard]==0.29.0

# HTTP Client
httpx==0.27.2

# ZMQ
pyzmq==26.4.0

# Task Queue
redis==5.0.4
celery==5.4.0

# Monitoring & Metrics
prometheus-client==0.21.0
prometheus-fastapi-instrumentator==7.0.0
psutil==6.1.0

# TOML parsing (Python 3.9 compatibility)
tomli==2.0.1
tomli-w==1.0.0
python-dotenv==1.0.1

# Google Assistant / OAuth
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-assistant-grpc==0.3.0
grpcio==1.66.1
protobuf==3.20.3

# AI/ML Models
openai-whisper==20240930
ultralytics==8.3.43
pillow==11.0.0
numpy==1.26.4
torch==2.5.1
piper-tts==1.2.0
ollama==0.4.4
mediapipe==0.10.11

# RAG (Knowledge Base)
chromadb==0.5.23
sentence-transformers==3.3.1

# Testing (optional - for development)
ruff==0.6.9
mypy==1.19.1
pytest==8.4.2
pytest-asyncio==0.24.0
pytest-timeout==2.4.0
